Introduce new module Edges for edge constraints
 - close #1
