+++ random-cd2pic.hs
@@ -2,6 +2,7 @@
 
 module Main (main) where
 
+import Edges
 import Types
 import Output
 
@@ -11,112 +12,34 @@ import Data.GraphViz
 
 import System.Random
 
-newtype Arrow = Arrow (String, String, Connection)
-type DiagramEdge = (String, String, Connection)
-
-connections :: Syntax -> [DiagramEdge]
-connections (is, as) =
-  [(s, e, Inheritance) | (s, Just e) <- is]
-  ++ [(s, e, Assoc t m1 m2 False) | (t, _, m1, s, e, m2) <- as]
-
-selfEdges :: [DiagramEdge] -> [DiagramEdge]
-selfEdges es = [x | x@(s, e, _) <- es, e == s]
-
-doubleConnections :: [DiagramEdge] -> [(DiagramEdge, DiagramEdge)]
-doubleConnections es =
-  [(x, y) | x@(s1, e1, _) <- es, s1 /= e1
-          , y@(s2, e2, _) <- filterFirst x es, s2 /= e2
-          , s1 == s2 && e1 == e2 || s1 == e2 && s2 == e1]
-
-multipleInheritances :: [DiagramEdge] -> [(DiagramEdge, DiagramEdge)]
-multipleInheritances es =
-  [(x, y) | x@(s1, e1, Inheritance) <- es, s1 /= e1
-          , y@(s2, e2, Inheritance) <- filterFirst x es, s2 /= e2
-          , s1 == s2 && e1 /= e2]
-
-inheritanceCycles :: [DiagramEdge] -> [[DiagramEdge]]
-inheritanceCycles = cycles isInheritance
-  where
-    isInheritance Inheritance = True
-    isInheritance _           = False
-
-compositionCycles :: [DiagramEdge] -> [[DiagramEdge]]
-compositionCycles = cycles isComposition
-  where
-    isComposition (Assoc Composition _ _ _) = True
-    isComposition _                         = False
-
-wrongLimits :: [DiagramEdge] -> [DiagramEdge]
-wrongLimits es =
-  [c | c@(_, _, Assoc t s@(sl, sh) e _) <- es
-     , isComposition t && (sh /= Just 1 || sl < 0 || sl > 1)
-       || not (inLimit s)
-       || not (inLimit e)]
-  where
-    isComposition Composition = True
-    isComposition _           = False
-    inLimit (l, Nothing)
-      | 0 <= l && l <= 2 = True
-      | otherwise        = False
-    inLimit (l, Just h)
-      | l == 0 && l <  h && h <= 2 = True
-      | l >  0 && l <= h && h <= 2 = True
-      | otherwise                  = False
-
-cycles :: (Connection -> Bool) -> [DiagramEdge] -> [[DiagramEdge]]
-cycles connectionFilter es =
-  [c:cs |   (s , e, cs) <- getPaths connectionFilter es
-        , c@(s', e', _) <- edges, s' == e, s == e']
-  where
-    edges               = filter (connectionFilter . connection) es
-    connection  (_,_,c) = c
-
-getPaths :: (Connection -> Bool) -> [DiagramEdge] -> [(String, String, [DiagramEdge])]
-getPaths connectionFilter es =
-  [path | c@(s, e, _) <- edges, s /= e, path <- getPath s e [c] edges]
-  where
-    edges               = filter (connectionFilter . connection) es
-    start       (s,_,_) = s
-    end         (_,e,_) = e
-    connection  (_,_,c) = c
-    getPath :: String
-            -> String
-            -> [DiagramEdge]  -- edges within the current path
-            -> [DiagramEdge]  -- still available edges
-            -> [(String, String, [DiagramEdge])]
-    getPath s e ps es' =
-      (s, e, ps) :
-      -- prevent cyclic paths
-      let es'' = filter ((s /=) . start) $ filter ((s /=) . end) es'
-      in [path | p@(s', e', _) <- es', s' /= e', s' /= e, s == e'
-               , path <- getPath s' e (p:ps) es'']
-
-filterFirst :: Eq a => a -> [a] -> [a]
-filterFirst _ []     = []
-filterFirst x (y:ys) = if x == y then ys else y : filterFirst x ys
-
 data Config = Config {
     classes      :: (Maybe Int, Maybe Int),
     aggregations :: (Maybe Int, Maybe Int),
     associations :: (Maybe Int, Maybe Int),
     compositions :: (Maybe Int, Maybe Int),
-    inheritances :: (Maybe Int, Maybe Int)
+    inheritances :: (Maybe Int, Maybe Int),
+    searchSpace  :: Int
   } deriving (Eq)
 
 main :: IO ()
 main =
-  generate (Config { classes = (Just 4, Just 4), aggregations = (Nothing, Nothing), associations = (Nothing, Nothing), compositions = (Nothing, Nothing), inheritances = (Nothing, Nothing) })
+  generate Config {
+      classes      = (Just 4, Just 4),
+      aggregations = (Nothing, Nothing),
+      associations = (Nothing, Nothing),
+      compositions = (Nothing, Nothing),
+      inheritances = (Nothing, Nothing),
+      searchSpace  = 10
+    }
   >>= drawCdFromSyntax "output" Pdf
-
-searchSpace = 10
 
 generate :: Config -> IO Syntax
 generate c = do
-  ncls <- oneOfFirst searchSpace $ toAvailable $ classes c
-  nins <- oneOfFirst searchSpace $ toAvailable $ inheritances c
-  ncos <- oneOfFirst searchSpace $ toAvailable $ compositions c
-  nass <- oneOfFirst searchSpace $ toAvailable $ associations c
-  nags <- oneOfFirst searchSpace $ toAvailable $ aggregations c
+  ncls <- oneOfFirst (searchSpace c) $ toAvailable $ classes c
+  nins <- oneOfFirst (searchSpace c) $ toAvailable $ inheritances c
+  ncos <- oneOfFirst (searchSpace c) $ toAvailable $ compositions c
+  nass <- oneOfFirst (searchSpace c) $ toAvailable $ associations c
+  nags <- oneOfFirst (searchSpace c) $ toAvailable $ aggregations c
   if isPossible ncls nins ncos nass nags
     then do
       es <- generateEdges (classs ncls) nins ncos nass nags
@@ -151,13 +74,13 @@ generate c = do
       | cla * (cla - 1) `div` 2 < inh + com + ass + agg = False
       | otherwise                                       = True
 
-generateEdges classes inh com ass agg = do
-  xs <- foldl (\es t -> es >>= flip generateEdge t) (return []) $
+generateEdges :: [String] -> Int -> Int -> Int -> Int -> IO [DiagramEdge]
+generateEdges classes inh com ass agg =
+  foldl (\es t -> es >>= flip generateEdge t) (return []) $
     replicate inh Nothing
     ++ replicate com (Just Composition)
     ++ replicate ass (Just Association)
     ++ replicate agg (Just Aggregation)
-  return xs
   where
     oneOf :: [a] -> IO a
     oneOf xs = do
@@ -210,7 +133,7 @@ minimise c = c {
       | x < y     = (Just $ x + 1, Just y)
       | otherwise = (Just x, Just y)
     decrease (Nothing, my     ) = (Just 0, my)
-    decrease (Just  x, Nothing) = (Just x, Just $ x + searchSpace)
+    decrease (Just  x, Nothing) = (Just x, Just $ x + searchSpace c)
     decrease (Just  x, Just  y)
       | x < y     = (Just x, Just $ y - 1)
       | otherwise = (Just x, Just y)
